refactor(tests): split endingGenerator tests into focused modules (#818)
Split endingGenerator.test.ts (500 lines) into:
- basic.test.ts (202 lines) - Core generation with required fields and tone
- advanced.test.ts (246 lines) - Play time, custom prompts, ending types
- errors.test.ts (176 lines) - Error handling and retry logic
- testHelpers.ts (164 lines) - Mock data factories and utilities

Benefits:
- All test files now under 300 lines (max is 246)
- Improved organization by feature category
- Easier to locate specific test scenarios
- Reduced duplication through shared helpers

Progress: 5 of 48 files refactored (10%)
