+++ src/lib/ai/__tests__/endingGenerator.testHelpers.ts
@@ -145,3 +145,20 @@ export function setupTestTimers() {
 export function cleanupTestTimers() {
   jest.useRealTimers();
 }
+
+/**
+ * Creates a standard mock ending response
+ */
+export function createMockEndingResponse(overrides?: {
+  epilogue?: string;
+  tone?: string;
+  achievements?: string[];
+}) {
+  return `{
+    "epilogue": "${overrides?.epilogue || 'The journey ends...'}",
+    "characterLegacy": "A true hero...",
+    "worldImpact": "Forever changed...",
+    "tone": "${overrides?.tone || 'triumphant'}",
+    "achievements": ${JSON.stringify(overrides?.achievements || ['Victory'])}
+  }`;
+}
